pyproject.toml
[tool.poetry]
name = "amman-bus-backend"
version = "0.1.0"
description = ""
authors = ["SARAH AHMAD HUDAIB <SARAHHUDAIB@YAHOO.COM>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
Markdown = "^3.3.7"
django-filter = "^22.1"
qrcode = "^7.3.1"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
